requirements.txt
# Core dependencies
sounddevice>=0.4.6
numpy>=1.21.0,<2.0.0  # Pin to NumPy 1.x for compatibility with faster-whisper
scipy>=1.7.0

# Transcription - flexible versions for compatibility
faster-whisper>=0.9.0

# Google API integration
google-api-python-client>=2.50.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.8.0

# Summarization (AI providers)
openai>=1.0.0
anthropic>=0.7.0

# Configuration and utilities
PyYAML>=5.4.0
python-dotenv>=0.19.0
APScheduler>=3.9.0

# System tray and UI
pystray>=0.19.0
Pillow>=8.0.0

# Logging and monitoring
rich>=12.0.0

# Web UI
flask>=2.3.0
waitress>=2.1.0  # Production WSGI server

# Development dependencies
pytest>=6.0.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0
black>=22.0.0
flake8>=4.0.0
pylint>=2.12.0
isort>=5.10.0
mypy>=0.950
pre-commit>=3.0.0
